+++ ExpandWorldEvents/EventLoader.cs
@@ -1,0 +1,76 @@
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using ExpandWorldData;
+using HarmonyLib;
+namespace ExpandWorld.Event;
+
+public class Loader
+{
+  public static Dictionary<RandomEvent, List<string>> EventToRequirentEnvironment = [];
+
+  public static RandomEvent FromData(Data data)
+  {
+    RandomEvent random = new()
+    {
+      m_name = data.name,
+      m_spawn = data.spawns.Select(Spawn.Loader.FromData).ToList(),
+      m_enabled = data.enabled,
+      m_random = data.random,
+      m_duration = data.duration,
+      m_nearBaseOnly = data.nearBaseOnly,
+      m_pauseIfNoPlayerInArea = data.pauseIfNoPlayerInArea,
+      m_biome = DataManager.ToBiomes(data.biome),
+      m_requiredGlobalKeys = DataManager.ToList(data.requiredGlobalKeys),
+      m_notRequiredGlobalKeys = DataManager.ToList(data.notRequiredGlobalKeys),
+      m_altRequiredPlayerKeysAny = DataManager.ToList(data.requiredPlayerKeys),
+      m_altNotRequiredPlayerKeys = DataManager.ToList(data.notRequiredPlayerKeys),
+      m_altRequiredKnownItems = DataManager.ToItemList(data.requiredKnownItems),
+      m_altRequiredNotKnownItems = DataManager.ToItemList(data.notRequiredKnownItems),
+      m_startMessage = data.startMessage,
+      m_endMessage = data.endMessage,
+      m_forceMusic = data.forceMusic,
+      m_forceEnvironment = data.forceEnvironment
+    };
+    EventToRequirentEnvironment[random] = DataManager.ToList(data.requiredEnvironments);
+    return random;
+  }
+  public static Data ToData(RandomEvent random)
+  {
+    Data data = new()
+    {
+      name = random.m_name,
+      spawns = random.m_spawn.Select(Spawn.Loader.ToData).ToArray(),
+      enabled = random.m_enabled,
+      random = random.m_random,
+      duration = random.m_duration,
+      nearBaseOnly = random.m_nearBaseOnly,
+      pauseIfNoPlayerInArea = random.m_pauseIfNoPlayerInArea,
+      biome = DataManager.FromBiomes(random.m_biome),
+      requiredGlobalKeys = DataManager.FromList(random.m_requiredGlobalKeys),
+      notRequiredGlobalKeys = DataManager.FromList(random.m_notRequiredGlobalKeys),
+      requiredPlayerKeys = DataManager.FromList(random.m_altRequiredPlayerKeysAny),
+      notRequiredPlayerKeys = DataManager.FromList(random.m_altNotRequiredPlayerKeys),
+      requiredKnownItems = DataManager.FromList(random.m_altRequiredKnownItems),
+      notRequiredKnownItems = DataManager.FromList(random.m_altRequiredNotKnownItems),
+      startMessage = random.m_startMessage,
+      endMessage = random.m_endMessage,
+      forceMusic = random.m_forceMusic,
+      forceEnvironment = random.m_forceEnvironment
+    };
+    return data;
+  }
+}
+
+
+[HarmonyPatch(typeof(RandEventSystem), nameof(RandEventSystem.SetRandomEvent)), HarmonyPriority(Priority.First)]
+public class MultipleConfigurationsPerEvent
+{
+  static void Prefix(RandEventSystem __instance, ref RandomEvent ev)
+  {
+    // Clients automatically get the first event by name.
+    // So this is for the server (self host) if multiple configurations are used.
+    ev = __instance.GetEvent(ev?.m_name);
+  }
+}